+++ crates/amaru-node/src/stages/mod.rs
@@ -15,4 +15,3 @@
 pub mod build_node;
 pub mod build_stage_graph;
 pub mod config;
-mod ledger;